🔨 refactor: 코드리뷰 수정
